CustomPropTypes -> PropTypes with named exports
